✨ Add source examples for deprecating commands
